launcher/requirements.txt
azure-functions==1.23.0
requests==2.32.5
